Add calculator and operate files
